extract closeParens
