+++ src/closeParens.js
@@ -1,0 +1,13 @@
+const closeParens = (result) => {
+  var numberOfOpenP = (result.match(/\(/g) || []).length;
+  var numberOfCloseP = (result.match(/\)/g) || []).length;
+  //this solution works but only if there is one parenthesis missing,
+  //instead, account for all parens missing
+
+  while (numberOfOpenP > numberOfCloseP) {
+    result = result + ")";
+    numberOfOpenP--;
+  }
+  return result;
+};
+export default closeParens;
